chore: fix type annotations for mypy compliance and apply black formatting

+++ mlfs/metrics/classification/accuracy_score.py
@@ -18,4 +18,4 @@ def accuracy_score(y_true: np.ndarray, y_pred: np.ndarray) -> float:
     if y_true.shape != y_pred.shape:
         raise ValueError("Shapes of y_true and y_pred must match")
 
-    return np.mean(y_true == y_pred)
+    return float(np.mean(y_true == y_pred))
